linalg: got rid of (almost) all compiler warnings, those remaining are 	probably due to bugs in the code (see FIXME's) sparse.h: defined specific type INT (default int32_t) for matrix elements 	  (setting INT to int64_t needs more debugging) linalg/Makefile: removed unused binaries in main target merge_linalg_sqrt.sh: default maxlevel now 6 (speeds up factorization of c59)
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@494 de4796e2-0b1e-0410-91b4-86bc23549317

+++ linalg/transpose.c
@@ -148,7 +148,7 @@ void rebuild()
                     ret += fscanf(fin_vchunks[i], "%u", &idx);
                     vec_uint_push_back(rows[k], idx);
                 }
-                BUG_ON(ret != nc);
+                BUG_ON(ret != (int) nc);
             }
         }
         for(k = 0 ; k < kmax ; k++) {
